requirements.txt
numpy
opencv-python-headless
dlib
face_recognition
scikit-learn
Pillow
Flask